Changes - v4
Changes:
Fixed bugs for startscene and pausescene, works as it should through out game.
Changed switch cases in introstate and ppintrostate that now goes back to the subscene when returning from pausestate and startstate instead of restarting the scene.
Added a "?" button that brings you to the control page when pressed.